Add info to decommissioned assets about who demoted the asset

+++ hyposoft/src/utils/decommissionutils.js
@@ -1,5 +1,6 @@
 import * as firebaseutils from './firebaseutils'
 import * as logutils from './logutils'
+import * as userutils from './userutils'
 
 function decommissionAsset(id,callback) {
     firebaseutils.assetRef.doc(id).get().then(doc => {
@@ -8,18 +9,24 @@ function decommissionAsset(id,callback) {
             return
         }
         const docData = doc.data()
-        firebaseutils.assetRef.doc(id).delete().then(() => {
-            logutils.addLog(id,logutils.ASSET(),logutils.DECOMMISSION(),docData)
-            firebaseutils.decommissionRef.add({...docData,timestamp: Date.now()}).then(() => callback(true))
-            .catch( error => {
-              console.log("Error getting documents: ", error)
+        firebaseutils.usersRef.doc(userutils.getLoggedInUser()).get().then(doc => {
+          if (!doc.exists) {
               callback(false)
-            })
-        })
-        .catch( error => {
-          console.log("Error getting documents: ", error)
-          callback(false)
-        })
+              return
+          }
+          firebaseutils.assetRef.doc(id).delete().then(() => {
+              logutils.addLog(id,logutils.ASSET(),logutils.DECOMMISSION(),docData)
+              firebaseutils.decommissionRef.add({...docData,timestamp: Date.now(),name: doc.data().username}).then(() => callback(true))
+              .catch( error => {
+                  console.log("Error getting documents: ", error)
+                  callback(false)
+              })
+              })
+          })
+          .catch( error => {
+            console.log("Error getting documents: ", error)
+            callback(false)
+          })
     })
     .catch( error => {
       console.log("Error getting documents: ", error)
